pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "glyph-animator"
version = "0.1.0"
description = "Font-to-Lottie animation pipeline using mathematically-grounded natural patterns"
requires-python = ">=3.10"
dependencies = [
    "fonttools>=4.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pillow>=9.0",
    "pre-commit>=3.0",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
